production/groq_requirements.txt
# Groq AI Integration Requirements for MCP Payments Server
# Production-ready dependencies for ultra-fast AI inference

# Core AI dependency
groq>=0.4.2

# HTTP client for async operations
httpx>=0.26.0

# FastAPI and dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Data validation and serialization
pydantic>=2.5.0

# JSON processing
orjson>=3.9.10

# Logging enhancements
structlog>=23.2.0

# Environment management
python-dotenv>=1.0.0

# Performance monitoring
prometheus-client>=0.19.0

# Type checking
typing-extensions>=4.8.0

# Error handling and retries
tenacity>=8.2.3

# Security
cryptography>=41.0.7

# Optional: For enhanced development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
mypy>=1.7.1 